Script for connecting files and database entries
Small script developed for the LHTT project to connect the files on the project's fileserver to the database entries. The script reads the fileserver and stores all files in a dictionary (full path + filename = key). Then, for every entry path and file name are squeezed through some regular expressions in order to look for anything it might be connected to (e.g. Archaeological Units AU, Finds, Tombs etc.). In some additional txt-files all file extensions and all folders which should not be taken into account in this process are written down. The result of the data iteration is then stored within the dictionary and stored in some txt-file. This txt-file can be loaded again when starting a new "Fileserver()".